♻️ refactor(dynamics): move legacy mockstream
