+++ src/galax/dynamics/_src/legacy/mockstream/__init__.py
@@ -1,0 +1,9 @@
+"""Mockstream.
+
+This is private API.
+
+"""
+
+__all__ = ["MockStreamGenerator"]
+
+from .mockstream_generator import MockStreamGenerator
