+++ src/galax/dynamics/_src/legacy/mockstream/_moving.py
