+++ src/galax/dynamics/_src/legacy/mockstream/df/base.py
