+++ src/galax/dynamics/_src/legacy/mockstream/df/chen24.py
