+++ src/galax/dynamics/_src/legacy/mockstream/df/fardal15.py
