+++ src/galax/dynamics/_src/legacy/mockstream/df/progenitor.py
