+++ src/galax/dynamics/_src/legacy/mockstream/mockstream_generator.py
@@ -17,12 +17,12 @@ from unxt.quantity import BareQuantity as FastQ
 
 import galax.coordinates as gc
 import galax.typing as gt
-from .arm import MockStreamArm
-from .core import MockStream
 from .df import AbstractStreamDF, ProgenitorMassCallable
 from .utils import cond_reverse
 from galax.dynamics._src.legacy.funcs import default_integrator, evaluate_orbit
 from galax.dynamics._src.legacy.integrator import Integrator
+from galax.dynamics._src.mockstream.arm import MockStreamArm
+from galax.dynamics._src.mockstream.core import MockStream
 from galax.dynamics._src.orbit import Orbit
 from galax.potential import AbstractPotential
