+++ src/galax/dynamics/_src/legacy/mockstream/utils.py
