+++ src/galax/dynamics/_src/mockstream/__init__.py
@@ -4,8 +4,10 @@ This is private API.
 
 """
 
-__all__ = ["MockStreamGenerator", "MockStreamArm", "MockStream"]
+__all__ = [
+    "MockStream",
+    "MockStreamArm",
+]
 
 from .arm import MockStreamArm
 from .core import MockStream
-from .mockstream_generator import MockStreamGenerator
