+++ src/galax/dynamics/mockstream.py
@@ -1,6 +1,7 @@
 """:mod:`galax.dynamics.mockstream`."""
 
 __all__ = [
+    # Legacy
     "MockStreamGenerator",
     "MockStreamArm",
     "MockStream",
@@ -16,14 +17,15 @@ from jaxtyping import install_import_hook
 from galax.setup_package import RUNTIME_TYPECHECKER
 
 with install_import_hook("galax.dynamics.dynamics", RUNTIME_TYPECHECKER):
-    from ._src.mockstream import MockStream, MockStreamArm, MockStreamGenerator
-    from ._src.mockstream.df import (
+    from ._src.legacy.mockstream import MockStreamGenerator
+    from ._src.legacy.mockstream.df import (
         AbstractStreamDF,
         ChenStreamDF,
         ConstantMassProtenitor,
         FardalStreamDF,
         ProgenitorMassCallable,
     )
+    from ._src.mockstream import MockStream, MockStreamArm
 
 # Cleanup
 del install_import_hook, RUNTIME_TYPECHECKER
